requirements.txt
gradio
numpy
fastapi
uvicorn
jinja2
Pillow
pandas